pyproject.toml
[project]
name = "aai-coursework-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openrouter>=0.9.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
]
